requirements.txt
aiohttp==4.0.0a1
alabaster==0.7.12
antlr4-python3-runtime==4.9.3
anyascii==0.3.1
asttokens==2.2.1
async-timeout==3.0.1
attrs==20.2.0
Babel==2.11.0
backcall==0.2.0
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
confection==0.0.4
contractions==0.1.73
cymem==2.0.7
datasets==2.2.1
decorator==5.1.1
dill==0.3.6
docformatter==1.5.0
docutils==0.15.2
emoji==2.2.0
exceptiongroup==1.0.4
executing==1.2.0
fairscale==0.4.12
filelock==3.8.1
flake8==6.0.0
flake8-bugbear==22.10.27
fsspec==2022.2.0
fuzzywuzzy==0.18.0
gitdb==4.0.10
gitdb2==4.0.2
GitPython==3.1.29
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
huggingface-hub==0.11.1
humanfriendly==10.0
hydra-core==1.3.0.dev1
idna==3.4
imagesize==1.4.1
importlib-metadata==4.2.0
importlib-resources==5.10.0
iniconfig==1.1.1
iopath==0.1.10
ipython==8.7.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonlines==3.1.0
langcodes==3.3.0
Markdown==3.3.2
markdown-it-py==0.5.8
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mock==4.0.3
multidict==4.7.6
multiprocess==0.70.14
murmurhash==1.0.9
myst-parser==0.12.10
ninja==1.10.2.4
nltk==3.7
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0.dev2
packaging==21.3
pandas==1.5.2
# parlai == 1.5.0
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pluggy==1.0.0
portalocker==2.6.0
preshed==3.0.8
prompt-toolkit==3.0.33
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-gfm==2.0.0
py-rouge==1.1
pyahocorasick==2.0.0b1
pyarrow==10.0.1
pyasn1==0.5.0rc2
pyasn1-modules==0.3.0rc1
pycodestyle==2.10.0
pydantic==1.10.5
pyflakes==3.0.1
Pygments==2.13.0
pyparsing==3.0.9
pyro-ppl==0.3.2
pytest==7.2.0
pytest-datadir==1.4.1
pytest-regressions==2.4.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq==25.0.0b1
regex==2022.10.31
requests==2.28.1
requests-mock==1.10.0
responses==0.18.0
rsa==4.9
sacremoses==0.0.53
scikit-learn==1.2.0rc1
scipy==1.9.3
sh==1.14.3
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
snowballstemmer==2.2.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
Sphinx==2.2.2
sphinx-autodoc-typehints==1.10.3
sphinx-rtd-theme==1.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==2.4.5
stack-data==0.6.2
subword-nmt==0.3.8
tensorboardX==2.5
textsearch==0.0.24
thinc==8.1.7
threadpoolctl==3.1.0
tokenizers==0.12.1
tomli==2.0.1
torch==1.13.1
torchtext==0.14.1
torchvision==0.14.1
tornado==6.2
tqdm==4.62.3
traitlets==5.6.0
transformers==4.18.0
typer==0.7.0
typing_extensions==4.4.0
Unidecode==1.3.6
untokenize==0.1.1
urllib3==1.26.13
wasabi==1.1.1
wcwidth==0.2.5
websocket-client==1.4.2
xxhash==3.1.0
yarl==1.8.2
zipp==3.11.0
